:recycle: refactor: PMD modification and pass on Ahievement function in AchievementScreen
